Stats corrigés
	* Modification de l'attaque speed and range a des valeurs entières

+++ trunk/AE-go_GameServer/src/com/aionemu/gameserver/model/gameobjects/stats/MonsterGameStats.java
@@ -31,7 +31,7 @@ public class MonsterGameStats extends NpcGameStats
 		super(npc,nst);
 	}
 	
-	private MonsterGameStats (Creature owner, int power, int health, int agility, int accuracy, int knowledge, int will, int mainHandAttack, int mainHandCritRate, int offHandAttack, int offHandCritRate, float attackSpeed, float attackRange) {
+	private MonsterGameStats (Creature owner, int power, int health, int agility, int accuracy, int knowledge, int will, int mainHandAttack, int mainHandCritRate, int offHandAttack, int offHandCritRate, int attackSpeed, int attackRange) {
 		super(null,power,health,agility,accuracy,knowledge,will,mainHandAttack,mainHandCritRate,offHandAttack,offHandCritRate,attackSpeed,attackRange);
 	}
 }
